feat(api): move logic in campaign-participation and add sharedParticipationCountLastThirtyDays property
