+++ api/src/prescription/campaign-participation/domain/usecases/get-campaign-participation-statistics.js
@@ -1,0 +1,8 @@
+const getCampaignParticipationStatistics = async ({
+  organizationId,
+  ownerId,
+  campaignParticipationStatisticsRepository,
+}) =>
+  await campaignParticipationStatisticsRepository.getParticipationCountOnPrescriberCampaigns(organizationId, ownerId);
+
+export { getCampaignParticipationStatistics };
