+++ api/src/prescription/campaign-participation/infrastructure/serializers/jsonapi/campaign-participation-statistics-serializer.js
@@ -3,8 +3,8 @@ import jsonapiSerializer from 'jsonapi-serializer';
 const { Serializer } = jsonapiSerializer;
 
 const serialize = function (statistics) {
-  return new Serializer('organization-participation-statistics', {
-    attributes: ['totalParticipationCount', 'completedParticipationCount'],
+  return new Serializer('campaign-participation-statistics', {
+    attributes: ['totalParticipationCount', 'completedParticipationCount', 'sharedParticipationCountLastThirtyDays'],
   }).serialize(statistics);
 };
