--- api/src/prescription/organization-learner/domain/usecases/get-organization-learner-statistics.js
@@ -1,7 +1,0 @@
-const getOrganizationLearnerStatistics = async ({ organizationId, ownerId, organizationLearnerStatisticsRepository }) =>
-  await organizationLearnerStatisticsRepository.getParticipationStatistics({
-    organizationId,
-    ownerId,
-  });
-
-export { getOrganizationLearnerStatistics };
