+++ api/src/prescription/organization-learner/domain/usecases/index.js
@@ -40,7 +40,6 @@ import * as analysisRepository from '../../infrastructure/repositories/analysis-
 import { repositories } from '../../infrastructure/repositories/index.js';
 import * as organizationLearnerActivityRepository from '../../infrastructure/repositories/organization-learner-activity-repository.js';
 import * as organizationLearnerFeatureRepository from '../../infrastructure/repositories/organization-learner-feature-repository.js';
-import * as organizationLearnerStatisticsRepository from '../../infrastructure/repositories/organization-learner-statistics-repository.js';
 import * as organizationParticipantRepository from '../../infrastructure/repositories/organization-participant-repository.js';
 import * as registrationOrganizationLearnerRepository from '../../infrastructure/repositories/registration-organization-learner-repository.js';
 import * as scoOrganizationParticipantRepository from '../../infrastructure/repositories/sco-organization-participant-repository.js';
@@ -86,9 +85,9 @@ const dependencies = {
   tokenService,
   passwordValidator,
   writeCsvUtils,
-  organizationLearnerStatisticsRepository,
 };
 
+import { getCampaignParticipationStatistics } from '../../../campaign-participation/domain/usecases/get-campaign-participation-statistics.js';
 import { createAndReconcileUserToOrganizationLearner } from './create-and-reconcile-user-to-organization-learner.js';
 import { createUserAndReconcileToOrganizationLearnerFromExternalUser } from './create-user-and-reconcile-to-organization-learner-from-external-user.js';
 import { findAssociationBetweenUserAndOrganizationLearner } from './find-association-between-user-and-organization-learner.js';
@@ -108,7 +107,6 @@ import { getAnalysisByTubes } from './get-analysis-by-tubes.js';
 import { getAttestationZipForDivisions } from './get-attestation-zip-for-divisions.js';
 import { getOrganizationLearner } from './get-organization-learner.js';
 import { getOrganizationLearnerActivity } from './get-organization-learner-activity.js';
-import { getOrganizationLearnerStatistics } from './get-organization-learner-statistics.js';
 import { getOrganizationLearnerWithParticipations } from './get-organization-learner-with-participations.js';
 import { getOrganizationToJoin } from './get-organization-to-join.js';
 import { updateOrganizationLearnerDependentUserPassword } from './update-organization-learner-dependent-user-password.js';
@@ -133,7 +131,7 @@ const usecasesWithoutInjectedDependencies = {
   getAttestationZipForDivisions,
   getOrganizationLearner,
   getOrganizationLearnerActivity,
-  getOrganizationLearnerStatistics,
+  getOrganizationLearnerStatistics: getCampaignParticipationStatistics,
   getOrganizationLearnerWithParticipations,
   getOrganizationToJoin,
   updateOrganizationLearnerDependentUserPassword,
